requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
aiohttp==3.7.4.post0
    # via
    #   -r requirements.in
    #   discord-py
    #   tweepy
async-lru==1.0.3
    # via tweepy
async-timeout==3.0.1
    # via aiohttp
attrs==22.1.0
    # via aiohttp
certifi==2022.6.15
    # via requests
chardet==4.0.0
    # via aiohttp
charset-normalizer==2.1.0
    # via requests
discord-py==1.7.3
    # via -r requirements.in
idna==3.3
    # via
    #   requests
    #   yarl
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
oauthlib==3.2.0
    # via
    #   requests-oauthlib
    #   tweepy
pydantic==1.9.1
    # via -r requirements.in
requests==2.28.1
    # via
    #   requests-oauthlib
    #   tweepy
requests-oauthlib==1.3.1
    # via tweepy
tweepy[async]==4.10.0
    # via -r requirements.in
typing-extensions==4.3.0
    # via
    #   aiohttp
    #   pydantic
urllib3==1.26.11
    # via requests
yarl==1.8.1
    # via aiohttp
